# docs/openapi.json
{
 "components": {
  "schemas": {
...
   "BookmarkIn": {
    "additionalProperties": false,
    "properties": {
     "actions": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "items": {},
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actions"
     },
     "check_method": {
      "$ref": "#/components/schemas/CheckMethod",
      "default": "auto"
     },
     "enabled": {
      "default": true,
      "title": "Enabled",
      "type": "boolean"
     },
     "fetch": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fetch"
     },
     "filter": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Filter"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "gate": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gate"
     },
     "highlight_mode": {
      "$ref": "#/components/schemas/HighlightMode",
      "default": "standard"
     },
     "info1": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info1"
     },
     "info2": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info2"
     },
     "info3": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info3"
     },
     "macro_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Macro Id"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 300,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "plugin": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plugin"
     },
     "priority": {
      "default": 0,
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Priority",
      "type": "integer"
     },
     "schedule": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Schedule"
     },
     "source_type": {
      "$ref": "#/components/schemas/SourceType",
      "default": "auto"
     },
     "url": {
      "minLength": 1,
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "url"
    ],
    "title": "BookmarkIn",
    "type": "object"
   },
   "BookmarkOut": {
    "additionalProperties": false,
    "properties": {
     "actions": {
      "$ref": "#/components/schemas/ActionsConfig"
     },
     "baseline_version_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseline Version Id"
     },
     "check_method": {
      "$ref": "#/components/schemas/CheckMethod"
     },
     "consecutive_errors": {
      "title": "Consecutive Errors",
      "type": "integer"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "current_interval_s": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Current Interval S"
     },
     "enabled": {
      "title": "Enabled",
      "type": "boolean"
     },
     "fetch": {
      "$ref": "#/components/schemas/FetchConfig"
     },
     "filter": {
      "$ref": "#/components/schemas/FilterConfig"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "gate": {
      "$ref": "#/components/schemas/GateConfig"
     },
     "gate_anchor_version_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gate Anchor Version Id"
     },
     "highlight_mode": {
      "$ref": "#/components/schemas/HighlightMode"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "info1": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info1"
     },
     "info2": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info2"
     },
     "info3": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info3"
     },
     "keyword_hits": {
      "items": {
       "type": "string"
      },
      "title": "Keyword Hits",
      "type": "array"
     },
     "last_changed_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Changed At"
     },
     "last_checked_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Checked At"
     },
     "latest_version_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latest Version Id"
     },
     "macro_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Macro Id"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "next_due_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Due At"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "overrides": {
      "additionalProperties": true,
      "title": "Overrides",
      "type": "object"
     },
     "plugin": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plugin"
     },
     "priority": {
      "title": "Priority",
      "type": "integer"
     },
     "schedule": {
      "$ref": "#/components/schemas/ScheduleConfig"
     },
     "source_type": {
      "$ref": "#/components/schemas/SourceType"
     },
     "status": {
      "$ref": "#/components/schemas/BookmarkStatus"
     },
     "unread": {
      "title": "Unread",
      "type": "boolean"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "id",
     "folder_id",
     "name",
     "url",
     "source_type",
     "check_method",
     "enabled",
     "priority",
     "schedule",
     "fetch",
     "filter",
     "gate",
     "highlight_mode",
     "actions",
     "overrides",
     "macro_id",
     "plugin",
     "info1",
     "info2",
     "info3",
     "note",
     "status",
     "unread",
     "consecutive_errors",
     "current_interval_s",
     "next_due_at",
     "last_checked_at",
     "last_changed_at",
     "latest_version_id",
     "baseline_version_id",
     "gate_anchor_version_id",
     "created_at",
     "updated_at"
    ],
    "title": "BookmarkOut",
    "type": "object"
   },
   "BookmarkPatch": {
    "additionalProperties": false,
    "description": "Partial update. Config dicts are merged into the stored overrides; ``null`` inside\none removes that override (the value falls back to the inherited default).",
    "properties": {
     "actions": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "items": {},
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Actions"
     },
     "check_method": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/CheckMethod"
       },
       {
        "type": "null"
       }
      ]
     },
     "enabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enabled"
     },
     "fetch": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fetch"
     },
     "filter": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Filter"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "gate": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gate"
     },
     "highlight_mode": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/HighlightMode"
       },
       {
        "type": "null"
       }
      ]
     },
     "info1": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info1"
     },
     "info2": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info2"
     },
     "info3": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Info3"
     },
     "macro_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Macro Id"
     },
     "move_to_root": {
      "default": false,
      "title": "Move To Root",
      "type": "boolean"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 300,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "plugin": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plugin"
     },
     "priority": {
      "anyOf": [
       {
        "maximum": 1.0,
        "minimum": 0.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Priority"
     },
     "schedule": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Schedule"
     },
     "source_type": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/SourceType"
       },
       {
        "type": "null"
       }
      ]
     },
     "url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Url"
     }
    },
    "title": "BookmarkPatch",
    "type": "object"
   },
...
   "BookmarkSummary": {
    "additionalProperties": false,
    "description": "The light row the bookmark list pages through (10,000 rows must stay fast).",
    "properties": {
     "check_method": {
      "$ref": "#/components/schemas/CheckMethod"
     },
     "consecutive_errors": {
      "title": "Consecutive Errors",
      "type": "integer"
     },
     "enabled": {
      "title": "Enabled",
      "type": "boolean"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "interval_s": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Interval S"
     },
     "keyword_hits": {
      "items": {
       "type": "string"
      },
      "title": "Keyword Hits",
      "type": "array"
     },
     "last_changed_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Changed At"
     },
     "last_checked_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Checked At"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "next_due_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Due At"
     },
     "priority": {
      "title": "Priority",
      "type": "integer"
     },
     "schedule_mode": {
      "$ref": "#/components/schemas/ScheduleMode"
     },
     "source_type": {
      "$ref": "#/components/schemas/SourceType"
     },
     "status": {
      "$ref": "#/components/schemas/BookmarkStatus"
     },
     "unread": {
      "title": "Unread",
      "type": "boolean"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "id",
     "folder_id",
     "name",
     "url",
     "source_type",
     "check_method",
     "enabled",
     "priority",
     "status",
     "unread",
     "consecutive_errors",
     "interval_s",
     "schedule_mode",
     "next_due_at",
     "last_checked_at",
     "last_changed_at"
    ],
    "title": "BookmarkSummary",
    "type": "object"
   },
   "BulkRequest": {
    "additionalProperties": false,
    "properties": {
     "action": {
      "enum": [
       "update",
       "move",
       "enable",
       "disable",
       "delete"
      ],
      "title": "Action",
      "type": "string"
     },
     "folder_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Folder Id"
     },
     "ids": {
      "items": {
       "type": "integer"
      },
      "minItems": 1,
      "title": "Ids",
      "type": "array"
     },
     "patch": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BookmarkPatch"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "required": [
     "ids",
     "action"
    ],
    "title": "BulkRequest",
    "type": "object"
   },
   "BulkResult": {
    "additionalProperties": false,
    "properties": {
     "affected": {
      "title": "Affected",
      "type": "integer"
     }
    },
    "required": [
     "affected"
    ],
    "title": "BulkResult",
    "type": "object"
   },
...
   "CheckMethod": {
    "enum": [
     "auto",
     "static",
     "browser",
     "screenshot"
    ],
    "title": "CheckMethod",
    "type": "string"
   },
   "CheckQueued": {
    "additionalProperties": false,
    "properties": {
     "queued": {
      "title": "Queued",
      "type": "integer"
     }
    },
    "required": [
     "queued"
    ],
    "title": "CheckQueued",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Page_BookmarkSummary_": {
    "additionalProperties": false,
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/BookmarkSummary"
      },
      "title": "Items",
      "type": "array"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     }
    },
    "required": [
     "items"
    ],
    "title": "Page[BookmarkSummary]",
    "type": "object"
   },
   "Page_ChangeOut_": {
    "additionalProperties": false,
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ChangeOut"
      },
      "title": "Items",
      "type": "array"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     }
    },
    "required": [
     "items"
    ],
    "title": "Page[ChangeOut]",
    "type": "object"
   },
   "Page_CheckRunOut_": {
    "additionalProperties": false,
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/CheckRunOut"
      },
      "title": "Items",
      "type": "array"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     }
    },
    "required": [
     "items"
    ],
    "title": "Page[CheckRunOut]",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "PageWatch engine",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/bookmarks": {
   "get": {
    "operationId": "list_bookmarks_bookmarks_get",
    "parameters": [
     {
      "in": "query",
      "name": "folder",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Folder"
      }
     },
     {
      "in": "query",
      "name": "subfolders",
      "required": false,
      "schema": {
       "default": true,
       "title": "Subfolders",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "unread",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Unread"
      }
     },
     {
      "in": "query",
      "name": "enabled",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Enabled"
      }
     },
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "in": "query",
      "name": "sort",
      "required": false,
      "schema": {
       "default": "id",
       "title": "Sort",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "desc",
      "required": false,
      "schema": {
       "default": false,
       "title": "Desc",
       "type": "boolean"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 500,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_BookmarkSummary_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Bookmarks",
    "tags": [
     "bookmarks"
    ]
   },
   "post": {
    "operationId": "create_bookmark_bookmarks_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BookmarkIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BookmarkOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Bookmark",
    "tags": [
     "bookmarks"
    ]
   }
  },
  "/bookmarks/bulk": {
   "post": {
    "operationId": "bulk_bookmarks_bulk_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BulkRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BulkResult"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Bulk",
    "tags": [
     "bookmarks"
    ]
   }
  },
  "/bookmarks/{bookmark_id}": {
   "delete": {
    "operationId": "delete_bookmark_bookmarks__bookmark_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "bookmark_id",
      "required": true,
      "schema": {
       "title": "Bookmark Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Bookmark",
    "tags": [
     "bookmarks"
    ]
   },
   "get": {
    "operationId": "get_bookmark_bookmarks__bookmark_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "bookmark_id",
      "required": true,
      "schema": {
       "title": "Bookmark Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BookmarkOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Bookmark",
    "tags": [
     "bookmarks"
    ]
   },
   "patch": {
    "operationId": "patch_bookmark_bookmarks__bookmark_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "bookmark_id",
      "required": true,
      "schema": {
       "title": "Bookmark Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BookmarkPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BookmarkOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Bookmark",
    "tags": [
     "bookmarks"
    ]
   }
  },
  "/bookmarks/{bookmark_id}/changes": {
   "get": {
    "operationId": "list_changes_bookmarks__bookmark_id__changes_get",
    "parameters": [
     {
      "in": "path",
      "name": "bookmark_id",
      "required": true,
      "schema": {
       "title": "Bookmark Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 500,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "before",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Before"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ChangeOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Changes",
    "tags": [
     "bookmarks"
    ]
   }
  },
  "/bookmarks/{bookmark_id}/check": {
   "post": {
    "operationId": "check_bookmark_bookmarks__bookmark_id__check_post",
    "parameters": [
     {
      "in": "path",
      "name": "bookmark_id",
      "required": true,
      "schema": {
       "title": "Bookmark Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "force",
      "required": false,
      "schema": {
       "default": false,
       "title": "Force",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CheckQueued"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Check Bookmark",
    "tags": [
     "bookmarks"
    ]
   }
  },
  "/bookmarks/{bookmark_id}/read": {
   "post": {
    "operationId": "mark_read_bookmarks__bookmark_id__read_post",
    "parameters": [
     {
      "in": "path",
      "name": "bookmark_id",
      "required": true,
      "schema": {
       "title": "Bookmark Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BookmarkOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Mark Read",
    "tags": [
     "bookmarks"
    ]
   }
  },
  "/bookmarks/{bookmark_id}/runs": {
   "get": {
    "operationId": "list_runs_bookmarks__bookmark_id__runs_get",
    "parameters": [
     {
      "in": "path",
      "name": "bookmark_id",
      "required": true,
      "schema": {
       "title": "Bookmark Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 500,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "before",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Before"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_CheckRunOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Runs",
    "tags": [
     "bookmarks"
    ]
   }
  },
...
 }
}